refactor: extract sub-components from Step2 and Step3
Step2: extract TemplateHeader (template name, tags, description block)
Step3: extract DockerImageSection (image+tag grid) and HardwareLimitsSection
       (CPU+memory grid), reducing Step3 from 132 to 70 lines
